fix(backend): MSE 公開 API の panic 排除とグリッドストライドの符号拡張
codex-review（PR #1078・イシュー #1045）の P1 指摘 2 件・Cursor Bugbot 指摘
1 件を修正する。

- CudaMse::run_mse_loss_f32／run_mse_backward_f32・MetalMse の同名メソッド
  は公開 API のため、pred/target 長さ不一致時の assert_eq!（panic）を
  CudaError::InvalidElementwiseShape／MetalError::InvalidElementwiseShape
  の型付きエラーへ置き換える（elementwise.rs::
  validate_elementwise_binary_dims と同じ構成の
  validate_mse_binary_len を追加）。ops.rs 経由でない直接呼び出しでも
  panic しない。
- kernels_mse.rs::MSE_PARTIAL_F32・shaders/mse.metal::mse_partial_f32 の
  grid-stride ループ添字を int/uint から long long／ulong へ拡張する。
  numel は i32::MAX／u32::MAX まで許容するため、int/uint のまま
  idx += stride を続けると numel 近傍で signed overflow／unsigned
  wraparound により idx < numel の境界チェックを迂回しうる
  （kernels_softmax.rs／softmax.metal で既に採用済みの対策と同じ方針）。

各修正に対応する回帰テスト（ソース文字列検査）を追加した。

Refs #1045

+++ crates/backend-metal/src/shaders/mse.metal
@@ -80,8 +80,15 @@ kernel void mse_partial_f32(
     threadgroup float simd_sums[MSE_SIMDGROUPS_PER_TG];
 
     float acc = 0.0f;
-    uint stride = grid_size * tg_size;
-    for (uint idx = tg_id * tg_size + tid; idx < numel; idx += stride) {
+    // REQ-8: `numel` は `u32::MAX`（`mse.rs::validate_mse_len`）まで
+    // 許容するため、`uint` 添字のまま `idx += stride` を続けると `numel`
+    // 近傍で unsigned wraparound（`idx` が 0 付近へ巻き戻る）により
+    // `idx < numel` の境界チェックを迂回しうる（Bugbot 指摘。CUDA 側
+    // `kernels_mse.rs::MSE_PARTIAL_F32` の `long long` 対応と同じ理由）。
+    // `softmax.metal`／`rmsnorm.metal` の `ulong row_base` と同じ方針で
+    // `ulong` を用いる。
+    ulong stride = (ulong)grid_size * (ulong)tg_size;
+    for (ulong idx = (ulong)tg_id * (ulong)tg_size + (ulong)tid; idx < numel; idx += stride) {
         float diff = pred[idx] - target[idx];
         acc = fma(diff, diff, acc);
     }
